+++ src/test/java/it/bigdatalab/algorithm/MHSEXTest.java
@@ -270,7 +270,7 @@ class MHSEXTest extends AlgoTest {
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("unCycleFarnessProvider")
-    void testFarness_UnCycle(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_UnCycle(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_undirected/32-cycle.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -290,13 +290,13 @@ class MHSEXTest extends AlgoTest {
         MHSEX algo = new MHSEX(g, param.getNumSeeds(), param.getThreshold(), nodes, param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("cycleFarnessProvider")
-    void testFarness_DiCycle(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_DiCycle(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_directed/32-cycle.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -316,13 +316,13 @@ class MHSEXTest extends AlgoTest {
         MultithreadExpansion algo = new MultithreadExpansion(g, param.getNumSeeds(), param.getThreshold(), nodes, param.getNumThreads(), param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("unWheelFarnessProvider")
-    void testFarness_UnWheel(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_UnWheel(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_undirected/32-wheel.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -342,13 +342,13 @@ class MHSEXTest extends AlgoTest {
         MHSEX algo = new MHSEX(g, param.getNumSeeds(), param.getThreshold(), nodes, param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("pathFarnessProvider")
-    void testFarness_DiPath(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_DiPath(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_directed/32-path.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -368,13 +368,13 @@ class MHSEXTest extends AlgoTest {
         MHSEX algo = new MHSEX(g, param.getNumSeeds(), param.getThreshold(), nodes, param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("tPathFarnessProvider")
-    void testFarness_DiTPath(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_DiTPath(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_directed/32t-path.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -394,13 +394,13 @@ class MHSEXTest extends AlgoTest {
         MHSEX algo = new MHSEX(g, param.getNumSeeds(), param.getThreshold(), nodes, param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("inStarFarnessProvider")
-    void testFarness_DiInStar(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_DiInStar(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_directed/32in-star.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -421,13 +421,13 @@ class MHSEXTest extends AlgoTest {
         MHSEX algo = new MHSEX(g, param.getNumSeeds(), param.getThreshold(), nodes, param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("outStarFarnessProvider")
-    void testFarness_DiOutStar(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_DiOutStar(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_directed/32out-star.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -447,7 +447,7 @@ class MHSEXTest extends AlgoTest {
         MHSEX algo = new MHSEX(g, param.getNumSeeds(), param.getThreshold(), nodes, param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 }
